Added packet type enum.
