+++ index.ts
@@ -1,15 +1,16 @@
-import aprsPacket from './src/aprsPacket';
-import aprsParser from './src/parser';
-import ConversionConstantEnum from './src/ConversionConstantEnum';
-import digipeater from './src/digipeater';
-import telemetry from './src/telemetry';
-import wx from './src/wx';
+import aprsPacket from './src/aprsPacket'
+import aprsParser from './src/parser'
+import { ConversionConstantEnum, PacketTypeEnum } from './src/enums'
+import digipeater from './src/digipeater'
+import telemetry from './src/telemetry'
+import wx from './src/wx'
 
 export {
-   aprsPacket
-   , aprsParser
-   , ConversionConstantEnum
-   , digipeater
-   , telemetry
-   , wx
-};
+    aprsPacket
+    , aprsParser
+    , ConversionConstantEnum
+    , digipeater
+    , PacketTypeEnum
+    , telemetry
+    , wx
+}
