+++ package.json
@@ -2,7 +2,7 @@
   "author": "afairhurst",
   "license": "ISC",
   "name": "js-aprs-fap",
-  "version": "1.1.1",
+  "version": "1.1.2-alpha",
   "homepage": "https://github.com/KD0NKS/js-aprs-fap",
   "description": "NodeJs library for parsing APRS packets.",
   "repository": {
