+++ src/enums/ConversionConstantEnum.ts
@@ -1,4 +1,4 @@
-const enum ConversionConstantEnum {
+export enum ConversionConstantEnum {
     KNOT_TO_KMH = 1.852         // nautical miles per hour to kilometers per hour
     , MPH_TO_KMH = 1.609344     // miles per hour to kilometers per hour
     , KMH_TO_MS = 10 / 36       // kilometers per hour to meters per second
@@ -6,5 +6,3 @@ const enum ConversionConstantEnum {
     , HINCH_TO_MM = 0.254       // hundredths of an inch to millimeters
     , FEET_TO_METERS = 0.3048
 }
-
-export default ConversionConstantEnum;
