+++ src/enums/PacketTypeEnum.ts
@@ -1,0 +1,12 @@
+export enum PacketTypeEnum {
+    CAPABILITIES = 'capabilities'
+    , DX = 'dx'
+    , ITEM = 'item'
+    , LOCATION = 'location'
+    , MESSAGE = 'message'
+    , OBJECT = 'object'
+    , STATUS = 'status'
+    , TELEMETRY = 'telemetry'
+    , TELEMETRY_MESSAGE = 'telemetry-message'
+    , WEATHER = 'wx'
+}
