+++ test/dxCluster.test.ts
@@ -1,0 +1,6 @@
+/*
+http://aprsisce.wikidot.com/kenwood-dx-cluster
+URCALL>RESORC:DX de YOURCALL.>11111111112222222222*33333333334444444444*55555
+URCALL>RESORC:DX de YOURCALL.>FIELD-ONE.ITEMNAME*FIELD-TWOFIELD-3333*FORTH
+URCALL>RESORC:DX de VE6SLP-10>144390.0 PIGEON 203 dg frm VE6SLP-10 74km
+*/
